Move common repository functions to shared repository component
