+++ src/Spd.Resource.Repository/BizLicApplication/BizLicApplicationRepository.cs
@@ -52,60 +52,21 @@ internal class BizLicApplicationRepository : IBizLicApplicationRepository
             app = _mapper.Map<spd_application>(cmd);
             _context.AddTospd_applications(app);
         }
-        LinkServiceType(cmd.WorkerLicenceTypeCode, app);
+        _context.LinkServiceType(cmd.WorkerLicenceTypeCode, app);
         if (cmd.HasExpiredLicence == true && cmd.ExpiredLicenceId != null) 
-            LinkExpiredLicence(cmd.ExpiredLicenceId, app);
+            _context.LinkExpiredLicence(cmd.ExpiredLicenceId, app);
         LinkOrganization(cmd.ApplicantId, app);
         LinkPrivateInvestigator(cmd.PrivateInvestigatorSwlInfo, app);
         await _context.SaveChangesAsync();
 
         //Associate of 1:N navigation property with Create of Update is not supported in CRM, so have to save first.
         //then update category.
-        ProcessCategories(cmd.CategoryCodes, app);
+        _context.ProcessCategories(cmd.CategoryCodes, app);
         await _context.SaveChangesAsync();
         return new BizLicApplicationCmdResp((Guid)app.spd_applicationid, cmd.ApplicantId);
     }
 
-    private void ProcessCategories(IEnumerable<WorkerCategoryTypeEnum> categories, spd_application app)
-    {
-        foreach (var c in categories)
-        {
-            var cat = _context.LookupLicenceCategory(c.ToString());
-            if (cat != null && !app.spd_application_spd_licencecategory.Any(c => c.spd_licencecategoryid == cat.spd_licencecategoryid))
-            {
-                _context.AddLink(app, nameof(spd_application.spd_application_spd_licencecategory), cat);
-            }
-        }
-        foreach (var appCategory in app.spd_application_spd_licencecategory)
-        {
-            var code = DynamicsContextLookupHelpers.LookupLicenceCategoryKey(appCategory.spd_licencecategoryid);
-            //if categories do not contain cat
-            if (!categories.Any(c => c.ToString() == code))
-            {
-                _context.DeleteLink(app, nameof(spd_application.spd_application_spd_licencecategory), appCategory);
-            }
-        }
-    }
-
-    private void LinkServiceType(WorkerLicenceTypeEnum? licenceType, spd_application app)
-    {
-        if (licenceType == null) throw new ArgumentException("invalid LicenceApplication type");
-        spd_servicetype? servicetype = _context.LookupServiceType(licenceType.ToString());
-        if (servicetype != null)
-        {
-            _context.SetLink(app, nameof(spd_application.spd_ServiceTypeId), servicetype);
-        }
-    }
-
-    private void LinkExpiredLicence(Guid? expiredLicenceId, spd_application app)
-    {
-        if (expiredLicenceId == null) return;
-        var licence = _context.spd_licences.Where(l => l.spd_licenceid == expiredLicenceId).FirstOrDefault();
-        if (licence != null)
-        {
-            _context.SetLink(app, nameof(spd_application.spd_CurrentExpiredLicenceId), licence);
-        }
-    }
+    
 
     private void LinkOrganization(Guid? accountId, spd_application app)
     {
