+++ src/Spd.Resource.Repository/LicenceApplication/LicenceApplicationRepository.cs
@@ -24,11 +24,12 @@ internal class LicenceApplicationRepository : ILicenceApplicationRepository
         spd_application app = _mapper.Map<spd_application>(cmd);
         app.statuscode = (int)ApplicationStatusOptionSet.Incomplete;
         _context.AddTospd_applications(app);
-        LinkServiceType(cmd.WorkerLicenceTypeCode, app);
+        _context.LinkServiceType(cmd.WorkerLicenceTypeCode, app);
         contact? contact = _mapper.Map<contact>(cmd);
         if (cmd.ApplicationTypeCode == ApplicationTypeEnum.New)
         {
-            if (cmd.HasExpiredLicence == true && cmd.ExpiredLicenceId != null) LinkExpiredLicence(cmd.ExpiredLicenceId, app);
+            if (cmd.HasExpiredLicence == true && cmd.ExpiredLicenceId != null) 
+                _context.LinkExpiredLicence(cmd.ExpiredLicenceId, app);
             //for new, always create a new contact
             contact = await _context.CreateContact(contact, null, _mapper.Map<IEnumerable<spd_alias>>(cmd.Aliases), ct);
         }
@@ -48,7 +49,7 @@ internal class LicenceApplicationRepository : ILicenceApplicationRepository
 
             if (cmd.OriginalLicenceId != null)
             {
-                LinkExpiredLicence(cmd.OriginalLicenceId, app);
+                _context.LinkExpiredLicence(cmd.OriginalLicenceId, app);
             }
             else
             {
@@ -62,7 +63,7 @@ internal class LicenceApplicationRepository : ILicenceApplicationRepository
         //then update category.
         if (cmd.WorkerLicenceTypeCode == WorkerLicenceTypeEnum.SecurityWorkerLicence)
         {
-            ProcessCategories(cmd.CategoryCodes, app);
+            _context.ProcessCategories(cmd.CategoryCodes, app);
         }
         await _context.SaveChangesAsync(ct);
         return new LicenceApplicationCmdResp((Guid)app.spd_applicationid, (Guid)contact.contactid);
@@ -117,13 +118,14 @@ internal class LicenceApplicationRepository : ILicenceApplicationRepository
                 _context.SetLink(app, nameof(spd_application.spd_ApplicantId_contact), contact);
             }
         }
-        LinkServiceType(cmd.WorkerLicenceTypeCode, app);
-        if (cmd.HasExpiredLicence == true && cmd.ExpiredLicenceId != null) LinkExpiredLicence(cmd.ExpiredLicenceId, app);
+        _context.LinkServiceType(cmd.WorkerLicenceTypeCode, app);
+        if (cmd.HasExpiredLicence == true && cmd.ExpiredLicenceId != null) 
+            _context.LinkExpiredLicence(cmd.ExpiredLicenceId, app);
         await LinkTeam(DynamicsConstants.Licensing_Client_Service_Team_Guid, app, ct);
         await _context.SaveChangesAsync();
         //Associate of 1:N navigation property with Create of Update is not supported in CRM, so have to save first.
         //then update category.
-        ProcessCategories(cmd.CategoryCodes, app);
+        _context.ProcessCategories(cmd.CategoryCodes, app);
         await _context.SaveChangesAsync();
         return new LicenceApplicationCmdResp((Guid)app.spd_applicationid, cmd.ApplicantId);
     }
@@ -165,46 +167,6 @@ internal class LicenceApplicationRepository : ILicenceApplicationRepository
         }
         return _mapper.Map<IList<LicenceAppListResp>>(applist.OrderByDescending(o => o.createdon));
 
-    }
-
-    private void ProcessCategories(IEnumerable<WorkerCategoryTypeEnum> categories, spd_application app)
-    {
-        foreach (var c in categories)
-        {
-            var cat = _context.LookupLicenceCategory(c.ToString());
-            if (cat != null && !app.spd_application_spd_licencecategory.Any(c => c.spd_licencecategoryid == cat.spd_licencecategoryid))
-            {
-                _context.AddLink(app, nameof(spd_application.spd_application_spd_licencecategory), cat);
-            }
-        }
-        foreach (var appCategory in app.spd_application_spd_licencecategory)
-        {
-            var code = DynamicsContextLookupHelpers.LookupLicenceCategoryKey(appCategory.spd_licencecategoryid);
-            //if categories do not contain cat
-            if (!categories.Any(c => c.ToString() == code))
-            {
-                _context.DeleteLink(app, nameof(spd_application.spd_application_spd_licencecategory), appCategory);
-            }
-        }
-    }
-    private void LinkServiceType(WorkerLicenceTypeEnum? licenceType, spd_application app)
-    {
-        if (licenceType == null) throw new ArgumentException("invalid LicenceApplication type");
-        spd_servicetype? servicetype = _context.LookupServiceType(licenceType.ToString());
-        if (servicetype != null)
-        {
-            _context.SetLink(app, nameof(spd_application.spd_ServiceTypeId), servicetype);
-        }
-    }
-
-    private void LinkExpiredLicence(Guid? expiredLicenceId, spd_application app)
-    {
-        if (expiredLicenceId == null) return;
-        var licence = _context.spd_licences.Where(l => l.spd_licenceid == expiredLicenceId).FirstOrDefault();
-        if (licence != null)
-        {
-            _context.SetLink(app, nameof(spd_application.spd_CurrentExpiredLicenceId), licence);
-        }
     }
 
     private async Task LinkTeam(string teamGuidStr, spd_application app, CancellationToken ct)
